Drift parameters are now defined in the setup file
The projection is now working

a new .geo.root

fixed padID                and adjust the output for the different macros

.geo file were missing

All the geometries are now in the tree format and a new class has been added: R3BGTPCElectronicsResponse. This class will be used to compare different electronics.

The electronics response has been added... It has still some problems that should be checked!

2 files deleted

The electronics response has been added... It has still some problems that should be checked!

Changed names from Fullv1 Fullv2 to FullBeamIn FullBeamOut. The target will be treated separately and it's in the passive folder. All the code has been cleaned. There is one problem with /sim/simHYDRA, it produces twice the GTPCPoint and this create problem with the lecture of the tree.

Changed names from Fullv1 Fullv2 to FullBeamIn FullBeamOut. The target will be treated separately and it's in the passive folder. A folder for the analysis has been produced. All the code has been cleaned.